[wpiutil,cscore] Move VideoMode.PixelFormat to wpiutil
This tracks the RawFrame move.
